refactor: move GamesSection to layout components directory
