+++ src/components/layout/GamesSection.tsx
